--- VladKornieiev/src/ACO/week2/homework/library/model/printed_editions/Book.java
@@ -1,11 +1,0 @@
-package ACO.week2.homework.library.model.printed_editions;
-
-/**
- * Created by v21k on 31.01.17.
- */
-public class Book extends PrintedEdition {
-
-    public Book(Genre genre, String name, String author, int year) {
-        super(genre, name, author, year);
-    }
-}
