--- VladKornieiev/src/ACO/week2/homework/library/model/printed_editions/Journal.java
@@ -1,11 +1,0 @@
-package ACO.week2.homework.library.model.printed_editions;
-
-/**
- * Created by v21k on 31.01.17.
- */
-public class Journal extends PrintedEdition {
-
-    public Journal(Genre genre, String name, String author, int year) {
-        super(genre, name, author, year);
-    }
-}
