+++ VladKornieiev/src/ACO/week2/homework/library/utils/Utils.java
@@ -1,29 +1,21 @@
 package ACO.week2.homework.library.utils;
 
 import ACO.data_structures.MyArrayList;
-import ACO.week2.homework.library.controller.interfaces.ILibraryReaderController;
-import ACO.week2.homework.library.model.Library;
-import ACO.week2.homework.library.model.Reader;
-import ACO.week2.homework.library.model.printed_editions.Book;
-import ACO.week2.homework.library.model.printed_editions.Genre;
-import ACO.week2.homework.library.model.printed_editions.Journal;
-import ACO.week2.homework.library.model.printed_editions.PrintedEdition;
+import ACO.week2.homework.library.model.*;
+import ACO.week2.homework.library.model.printed_editions.*;
 
 import java.util.Arrays;
 import java.util.Comparator;
-import java.util.Scanner;
 
 /**
  * Created by v21k on 31.01.17.
  */
 public class Utils {
 
-    private static Scanner sc = new Scanner(System.in);
-
     // to make our library alive
     public static boolean initLibrary(Library library) {
-        Journal journal = new Journal(Genre.JOURNAL, "Vogue", "Some Author", 2015);
-        Book book = new Book(Genre.FANTASY, "Harry Potter", "Joanne Rowling", 2008);
+        PrintedEdition journal = new PrintedEdition(Genre.JOURNAL, "Vogue", "Some Author", 2015);
+        PrintedEdition book = new PrintedEdition(Genre.FANTASY, "Harry Potter", "Joanne Rowling", 2008);
         PrintedEdition printedEdition = new PrintedEdition(Genre.SCIENCE, "Science", "MTI", 2013);
 
         library.addPrintedEdition(journal);
@@ -52,7 +44,9 @@ public class Utils {
 
         StringBuilder sb = new StringBuilder();
         for (int i = 0; i < printedEditions.length; i++) {
-            sb.append(i + 1).append(": ").append(printedEditions[i].toString()).append("\n");
+            if (printedEditions[i] != null) {
+                sb.append(i + 1).append(": ").append(printedEditions[i].toString()).append("\n");
+            }
         }
         return sb.toString();
     }
@@ -68,20 +62,6 @@ public class Utils {
             sb.append(i + 1).append(": ").append(al.get(i).toString()).append("\n");
         }
         return sb.toString();
-    }
-
-    public static int getComparatorChoice() {
-        System.out.println("Choose sort comparator: ");
-        System.out.println("1. By name");
-        System.out.println("2. By author");
-        System.out.println("3. By year");
-        return sc.nextInt();
-    }
-
-    public static int getReaderChoice(ILibraryReaderController lr) {
-        System.out.println("Choose reader (number)");
-        System.out.println(lr.showAllReaders());
-        return sc.nextInt();
     }
 
     public static Reader[] readersToArray(MyArrayList myArrayList) {
