+++ VladKornieiev/src/ACO/week2/homework/library/view/menu/MenuPrintedEditions.java
@@ -1,14 +1,9 @@
 package ACO.week2.homework.library.view.menu;
 
-import ACO.week2.homework.library.comparators.PEByAuthorComparator;
-import ACO.week2.homework.library.comparators.PEByNameComparator;
-import ACO.week2.homework.library.comparators.PEByYearComparator;
-import ACO.week2.homework.library.controller.interfaces.ILibraryPEController;
-import ACO.week2.homework.library.controller.interfaces.ILibraryReaderController;
-import ACO.week2.homework.library.model.Library;
-import ACO.week2.homework.library.model.Reader;
-import ACO.week2.homework.library.model.printed_editions.Genre;
-import ACO.week2.homework.library.model.printed_editions.PrintedEdition;
+import ACO.week2.homework.library.comparators.*;
+import ACO.week2.homework.library.controller.interfaces.*;
+import ACO.week2.homework.library.model.*;
+import ACO.week2.homework.library.model.printed_editions.*;
 import ACO.week2.homework.library.utils.Utils;
 
 import java.util.Comparator;
@@ -25,7 +20,7 @@ public class MenuPrintedEditions {
     // 2
     public static boolean showAllPE(ILibraryPEController lp) {
 
-        int choice = Utils.getComparatorChoice();
+        int choice = MenuUtils.getComparatorChoice();
 
         if (choice < 1 || choice > comparators.length) {
             // default
@@ -64,7 +59,7 @@ public class MenuPrintedEditions {
         System.out.println(Utils.showPrintedEditions(library.getPrintedEditionsInLibrary()));
         int choicePE = sc.nextInt();
 
-        int choiceReader = Utils.getReaderChoice(lr);
+        int choiceReader = MenuUtils.getReaderChoice(lr);
 
         if (choicePE >= 1 && choicePE <= library.getPrintedEditionsInLibrary().size()
                 && choiceReader >= 1 && choiceReader <= library.getReaders().size()) {
@@ -79,7 +74,7 @@ public class MenuPrintedEditions {
     // 6
     public static boolean showPEReaderTook(ILibraryPEController lp) {
 
-        int choice = Utils.getComparatorChoice();
+        int choice = MenuUtils.getComparatorChoice();
 
         if (choice < 1 || choice > comparators.length) {
             // default
@@ -93,7 +88,7 @@ public class MenuPrintedEditions {
     // 7
     public static boolean showCertainReaderPEs(Library library, ILibraryPEController lp, ILibraryReaderController lr) {
 
-        int choice = Utils.getReaderChoice(lr);
+        int choice = MenuUtils.getReaderChoice(lr);
 
         if (choice >= 1 && choice <= library.getReaders().size()) {
             System.out.println(lp.showPrintedEditionCertainReader((Reader) library.getReaders().get(choice - 1),
