+++ VladKornieiev/src/ACO/week2/homework/library/view/menu/MenuReader.java
@@ -1,9 +1,7 @@
 package ACO.week2.homework.library.view.menu;
 
 import ACO.week2.homework.library.controller.interfaces.ILibraryReaderController;
-import ACO.week2.homework.library.model.Library;
-import ACO.week2.homework.library.model.Reader;
-import ACO.week2.homework.library.utils.Utils;
+import ACO.week2.homework.library.model.*;
 
 import java.util.Scanner;
 
@@ -33,7 +31,7 @@ public class MenuReader {
 
     // 8
     public static boolean addReaderToBlackList(Library library, ILibraryReaderController lr) {
-        int choice = Utils.getReaderChoice(lr);
+        int choice = MenuUtils.getReaderChoice(lr);
         if (choice >= 1 && choice <= library.getReaders().size()) {
             System.out.println("Done!");
             return lr.addReaderToBlackList(((Reader) library.getReaders().get(choice - 1)));
